[Stats] Fixed bug with stats
I noticed that when there are no iw4x.stat at the mod specific folder and you try to play the game (/devmap <map>) you would get stuck on menu "popup_gettingdata".

Now, if the game is trying to read the stat and it doesn't exist, the game will call a function (?) idk what the function does exactly, but with it I have noticed the game will save iw4x.stat whenever the game is exited, map is launched and mod is loaded.

I have tested around for a bit now and everything should work now.